[#20] Minor Cleanup before pull request.
Removed the element from the index file. It is still included so you can
add it from the command line.

There still needs to be some discussion about how exactly we want to use
traceur and how we want to go about loading the compiled files. For now
System.js seems to be working fine.