[#18] Task and Server files pass jscs settings.
All the files in tasks/ and server/ are up to snuff with the style
settings. Gulpfile, config.paths.js and karma.conf.js are also properly
formatted.

Updated jscs task file with a bunch of new jscs tasks:
  - jscs:client
  - jscs:tasks
  - jscs:tests
  - jscs:server
  - jscs:root
  - jscs:all
  - jscs <-- alias for "jscs:all" right now
  - jscs:watch <-- watches all js files, ie "jscs:watch"

JSCS task is now run as part of the dev task (currently breaks build)

TODO
  - Fix jscs task as part of dev build to not break
  - Make sure the jscs:watch task works properly