[#79] Refactored some of the code.
The modal is now working - in compliance with the spec. But the console.log function in the example usage is coming up undefined. Tried console logging for the other components and same undefined error is showing up. Not sure what the problem is. It works when inputting it directly into the console.