[#303] Updated eslintrc with new rules
There have been a few updates to eslint so I updated the settings and
rules, removing deprecated ones and adding new ones. I also created a
new eslintrc file for the domain folder as the ES6 settings were
conflicting with the strict settings in non-es6 code.

Also did some clean up of errors that are easy to fix