[#432] Object stores and indexes are built from config object.
This should solve the bug with promised db, but in it's current state it
will fail hard when trying to upgrade a database to a new version. If
the database does not exist multiple config objects will be visited and
the db will be created correctly. If the db exists but is a version
behind an error occurs as the init process tries to recreate object
stores that already exist.