[#487] Worked on referrals remaining issue.
Once profile/get is created, then I will be able to test it. But it looks like it can grab the number of referrals left and print it onto the view.
Changed the registration-view.js file so that it creates the appropriate type of user.
Changed data binding for the referral remaining number.
Needed to change the regex validation to include ability to create email addresses with + in them.