[#509] Finished Analytics service and events.
Also did some cleanup in some places, modelType is now a prototype
member with a lookup on the instance's constructor. Although the
"modelType" still needs to be added to the arguments when stored in
IndexedDB so that the transformer function can still find the correct
constructor.

Artist profile now refreshes the model when the route changes.