Created EDChart and added some things to EDCollection
EDChart is a thing now, doesn't do much ATM.
EDCollection now has a third parameter for many of it's functions, a
boolean value which will make the function return a single Promise
instead of an array of Promises.
Also added to EDCollection is the "getInSequence" function which will
get a range of data but only one at a time instead of all at once.