Fixed some url issues.
Added a "dynamic" base url tag, updated ed-icon and img-placeholder js
to use correct url.

URLs in LESS files still seem to be an issue. The background image url
wasn't working so I tried a few things. I think we will need to refactor
the background image mixin to take the url as it would be relative from
the LESS file requiring it. If the only place the background image will
be used is view files then the current hack will work.

EDTrack is having some loading issues, not sure how to go about fixing
it as all of the other object files load fine.