Fixed some login weirdness.
I had changed the fall back route to be the discover page but that was
causing a problem where the app was going to the discover page when
there was no open session. I changed the default page back to the login
page. Then added a little logic to the login page attached callback to
redirect to discover if there is already an open session.