Profile pages have badges
Also moved some of the session restore logic into a function of the user
service. As well as stop redirecting on session restore.

Help resolve the flickering login page I moved the session restore stuff
into a new script "load.script.js" which is run before elements are
registered with polymer and the application runs.