Profile/Media Observation added to data service.
There were a few places where data needed to be updated on update.
Exposing observation of items in the LRU by the data service was
implemented.

The user service now observes the currently logged in user profile for
changes, so when the profile data updates the userService.currentProfile
gets updated and doesn't hang on to old references.

Edit profile now watches for changes so it can update the profile
picture on change.
The main side menu watches for changes to the logged in user so it can
update it's thumbnail as well.

Note: it was found that the edit profile API route returns a profile
object that is inconsistent with the profile get API route. This will be
added to the land mines document as a client side fix is in place for now.